+++ koin-projects/koin-android-viewmodel/src/main/java/org/koin/android/viewmodel/compat/ScopeCompat.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2020 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,6 +27,7 @@ import org.koin.core.scope.Scope
  * Scope functions to help for ViewModel in Java
  *
  * @author Jeziel Lago
+ * @author Arnaud Giuliani
  */
 object ScopeCompat {