+++ koin-projects/koin-androidx-scope/src/main/java/org/koin/androidx/scope/ScopeActivity.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2020 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.koin.androidx.scope
 
 import android.os.Bundle