+++ koin-projects/koin-androidx-workmanager/src/main/java/org/koin/androidx/workmanager/dsl/ScopeDSLExt.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2020 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,6 +25,8 @@ import org.koin.dsl.ScopeDSL
 import org.koin.dsl.bind
 
 /**
+ * Declare a scoped worker
+ *
  * @author Arnaud Giuliani
  **/
 inline fun <reified T : ListenableWorker> ScopeDSL.worker(