+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/context/KoinContext.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2020 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ import org.koin.core.Koin
 import org.koin.core.KoinApplication
 
 /**
- * Component thath hold the Koin instance
+ * Hold Current Koin context
  */
 interface KoinContext {