+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/error/BadScopeInstanceException.kt
@@ -1,3 +1,0 @@
-package org.koin.core.error
-
-class BadScopeInstanceException(s: String) : Exception(s)