+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/error/NoBeanDefFoundException.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2020 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,6 @@
 package org.koin.core.error
 
 /**
- * Definition
+ * Definition not Found
  */
 class NoBeanDefFoundException(msg: String) : Exception(msg)