+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/registry/InstanceRegistry.kt
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2017-2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.koin.core.registry
 
 import org.koin.core.Koin
@@ -18,7 +33,6 @@ import kotlin.reflect.KClass
 @OptIn(KoinInternal::class)
 class InstanceRegistry(val _koin: Koin, val _scope: Scope) {
 
-    //TODO Lock - ConcurrentHashMap
     private val _instances = HashMap<IndexKey, InstanceFactory<*>>()
     val instances: Map<IndexKey, InstanceFactory<*>>
         get() = _instances