+++ koin-projects/koin-ktor/src/main/kotlin/org/koin/ktor/ext/KoinApplicationCallExt.kt
@@ -1,7 +1,5 @@
-package org.koin.ktor.ext
-
 /*
- * Copyright 2017-2020 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +13,7 @@ package org.koin.ktor.ext
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.koin.ktor.ext
 
 import io.ktor.application.*
 import org.koin.core.context.GlobalContext